Quiz game v1.0
This is the very first version of this project which I wrote a couple of days ago. This version is not as functional as I plan to build the future versions, but I use this as some sort of concept of what the future versions of this project should do in a much better and advanced way. I will also introduce the MySQL database to the code in the next commit.