Quiz game with MySQL
In this project, I've created a very complex program with login features, Accounts, creating and rating quizzes, and so on... With MySQL implementation, I've succeeded to create a database that holds data for quizzes, questions, answers, and users. Within the program, the data in the database is being used and edited throughout the program to create the best user experience. The program itself is less than 500 lines long as it holds a lot of features. It is also not hard to use. After creating this, I've learned how to use MySQL and I hope it will be useful sometime.