Add files via upload
I've started implementing the login/logout feature into the app. There are a few minor bugs that need to be fixed but overall this feature works well, with a lot less code and struggle while using the app than before. In this feature, the user has an option to log in and stay logged in and use the app on his behalf as much as he wants until he decides to log out. Therefore he doesn't need to log in again to create quiz or quiz questions or anything like that. After he is finished using the program, another user can come in, log into his account and use the program until he logs out.