Quiz App v2.2
Introducing Manual Mode. Manual is basically some type of eraser for all the mistakes created in the database, but it also holds the ability to control the database. As of now, it's for developers only but later these features will also be implemented into the program (with some modifications) to be user friendly and maximize features. 

An idea for Manual Mode was created when I accidentally made a type in MySQL query and an entire element of the database didn't get upgraded. So now I can fix those development mistakes while also creating a feature for the users which will be implemented in the future. 

For now, only the quiz name alteration is added as a Manual Mode option, but soon, everything will be customizable.